More refactoring to move appropriate files/folders under pkg and web.
